requirements.txt
pytest-selenium==1.14.0